Refactor Sidebar into its own feature folder
