+++ src/features/Status/StatusPage.tsx
@@ -1,7 +1,7 @@
 import { Flex } from '@rebass/grid';
 import * as React from 'react';
 import { BackgroundImage, Button, H1, Paragraph } from '../../shared/Elements';
-import SideBarComponent from './Sidebar';
+import Sidebar from '../Sidebar/Sidebar';
 
 import Background from '../../assets/images/statuspage-background.svg';
 
@@ -9,7 +9,7 @@ class StatusPage extends React.Component<{}, {}> {
   public render() {
     return (
       <Flex flexDirection={'column'} alignItems={'center'}>
-        <SideBarComponent currentPage={'Home'} />
+        <Sidebar currentPage={'Home'} />
         <Flex flexDirection={'rows'} style={{ marginTop: '8em' }}>
           <H1 textAlign={'center'} display={'inline'}>
             Status:
