+++ src/shared/Elements/index.ts
@@ -12,5 +12,3 @@ export * from './Table';
 export * from './Textarea';
 export * from './ViewPDF';
 export * from './LinkDuo';
-export * from './Sidebar';
-export * from './SidebarMenuItem';
